+++ app/instance-initializers/axe-component.js
@@ -1,12 +1,14 @@
 import { assert } from '@ember/debug';
 import Component from '@ember/component';
 import { computed } from '@ember/object';
-import { isEmpty } from '@ember/utils';
+import { isEmpty, isPresent } from '@ember/utils';
 import { isArray } from '@ember/array';
 import { scheduleOnce } from '@ember/runloop';
 import Ember from 'ember';
 import ENV from '../config/environment';
 import isBackgroundReplacedElement from 'ember-a11y-testing/utils/is-background-replaced-element';
+import formatViolation from 'ember-a11y-testing/utils/format-violation';
+import violationsHelper from 'ember-a11y-testing/utils/violations-helper';
 
 const VIOLATION_CLASS__LEVEL_1 = 'axe-violation--level-1';
 const VIOLATION_CLASS__LEVEL_2 = 'axe-violation--level-2';
@@ -40,12 +42,16 @@ export function initialize() {
   const { componentOptions = {} } = addonConfig;
 
   const {
-    turnAuditOff,
+    turnAuditOff: configuredTurnAuditOff,
     axeOptions,
     axeCallback,
-    visualNoiseLevel,
-    axeViolationClassNames
+    visualNoiseLevel: configuredVisualNoiseLevel,
+    axeViolationClassNames: configuredAxeViolationClassNames
   } = componentOptions;
+
+  let visualNoiseLevel = configuredVisualNoiseLevel ? configuredVisualNoiseLevel : 1;
+  let axeViolationClassNames = isPresent(configuredAxeViolationClassNames) ? configuredAxeViolationClassNames : [];
+  let turnAuditOff = configuredTurnAuditOff || false;
 
   Component.reopen({
     /**
@@ -72,7 +78,7 @@ export function initialize() {
      * @default false
      * @type {Boolean}
      */
-    turnAuditOff: turnAuditOff || false,
+    turnAuditOff,
 
     /**
      * An array of classNames (or a space-separated string) to add to the component when a violation occurs.
@@ -83,7 +89,7 @@ export function initialize() {
      * @type {(Array|string)}
      * @see(https://github.com/ember-a11y/ember-a11y-testing/blob/master/content-for/head-footer.html)
      */
-    axeViolationClassNames: [],
+    axeViolationClassNames,
 
     /**
      * A numeric setting to determine the class applied to elements with violations
@@ -92,7 +98,7 @@ export function initialize() {
      * @type {string}
      * @see(https://github.com/ember-a11y/ember-a11y-testing/blob/master/content-for/head-footer.html)
      */
-    visualNoiseLevel: 1,
+    visualNoiseLevel,
 
     /**
      * Computes class name to be set on the element according to the
@@ -153,15 +159,15 @@ export function initialize() {
             nodes = violation.nodes;
 
             if (isEmpty(nodes) || nodes.length === 0) {
-              Ember.Logger.error(`[${violation.impact}]: ${violation.help} \nOffending markup is: \n \n${violation.helpUrl}`, violation);
-              window.violationsHelper.push(violation);
+              Ember.Logger.error(formatViolation(violation), violation);
+              violationsHelper.push(violation);
             }
 
             for (let j = 0, k = nodes.length; j < k; j++) {
               nodeData = nodes[j];
 
-              Ember.Logger.error(`[${violation.impact}]: ${violation.help} \nOffending markup is: \n${nodeData.html} \n${violation.helpUrl}`, violation);
-              window.violationsHelper.push(violation);
+              Ember.Logger.error(formatViolation(violation, nodeData.html), violation);
+              violationsHelper.push(violation);
 
               if (nodeData) {
                 nodeElem = document.querySelector(nodeData.target.join(','));
@@ -181,7 +187,7 @@ export function initialize() {
             this.axeCallback(results);
           }
 
-          scheduleOnce('afterRender', window.violationsHelper, window.violationsHelper.logTip);
+          scheduleOnce('afterRender', violationsHelper, violationsHelper.logTip);
         });
       }
     },